# src/CardLedger.Application/DTOs/PurchaseDtos.cs
namespace CardLedger.Application.DTOs;

public sealed record PurchaseRequest(
    string CardNumber,
    string ExpiryDate,
    string Cvv,
    decimal Amount,
    string Currency,
    string Description);

public sealed record PurchaseResponse(
    Guid Id,
    decimal Amount,
    string Currency,
    string Description);